Readable Textboxes, Plant Facts for Natives, and Grab/Forest Sound Effects
I added some sounds that I found. I realized I could use the same script for grabbing spray bottle and axe, like holding the invasive plants.

I also edited the GrabbableObject.cs file that was given to be able to remove those text boxes.... (it was a lot of work researching how to do this but did it!)

I'm having trouble how to make the Native plants grabbable but I included the plant facts for them.